requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-login>=0.6
flask-mail>=0.10
flask-wtf>=1.2  # CSRF protection
flask-limiter>=3.5  # Rate limiting
werkzeug>=3.0
python-dotenv>=1.0
requests>=2.31
bleach>=6.1  # Input sanitization
pyotp>=2.9
qrcode[pil]>=7.4
Pillow>=10.0
reportlab>=4.0
python-barcode>=0.15
apscheduler>=3.10
google-auth>=2.28
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
pytz>=2024.1
Authlib>=1.3
gunicorn==21.2.0